backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.5
firebase-admin==6.5.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.11.1
openai==2.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.15.0
uvicorn==0.32.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
